chore: update repo file explorer query, parsers, types and import paths
