+++ src/components/modules/RepoFileExplorer/FileExplorerHeader.tsx
@@ -1,7 +1,7 @@
 import Link from '@components/atoms/Link';
 import { getPastDate } from '@utils/date-formatter';
+import { CommitSummary } from '@utils/parsers/types/FileExplorer';
 import Image from 'next/image';
-import { CommitSummary } from '../RepoHomeFileExplorer/parse-file-explorer';
 
 type FileExplorerHeaderProps = {
   summary: Partial<CommitSummary>;
