+++ src/components/modules/RepoHomeFileExplorer/FileExplorerLineItem.tsx
@@ -1,7 +1,7 @@
 import Icon from '@components/atoms/Icon';
 import Link from '@components/atoms/Link';
 import { getPastDate } from '@utils/date-formatter';
-import { ExplorerLineItemCommitHistory } from './parse-file-explorer';
+import { ExplorerLineItemCommitHistory } from '@utils/parsers/types/FileExplorer';
 
 type FileExplorerLineItemProps = {
   owner: string;
